refactor home to dashboard
