+++ src/app/pages/dashboard/dashboard.module.ts
@@ -2,17 +2,17 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { IonicModule } from '@ionic/angular';
 import { FormsModule } from '@angular/forms';
-import { HomePage } from './home.page';
+import { DashboardPage } from './dashboard.page';
 
-import { HomePageRoutingModule } from './home-routing.module';
+import { DashboardPageRoutingModule } from './dashboard-routing.module';
 
 @NgModule({
   imports: [
     CommonModule,
     FormsModule,
     IonicModule,
-    HomePageRoutingModule,
+    DashboardPageRoutingModule,
   ],
-  declarations: [HomePage],
+  declarations: [DashboardPage],
 })
-export class HomePageModule {}
+export class DashboardPageModule {}
