+++ src/app/pages/dashboard/dashboard.page.spec.ts
@@ -5,23 +5,23 @@ import { RouterTestingModule } from '@angular/router/testing';
 import { Router } from '@angular/router';
 import { configureTestSuite } from 'ng-bullet';
 import { StoreModule } from '@ngrx/store';
-import { HomePage } from './home.page';
-import { AuthenticationProvider } from '../providers/authentication/authentication';
-import { appInfoReducer } from '../../store/app-info/app-info.reducer';
-import { AuthenticationProviderMock } from '../providers/authentication/__mocks__/authentication.mock';
+import { DashboardPage } from './dashboard.page';
+import { AuthenticationProvider } from '../../providers/authentication/authentication';
+import { appInfoReducer } from '../../../store/app-info/app-info.reducer';
+import { AuthenticationProviderMock } from '../../providers/authentication/__mocks__/authentication.mock';
 
-describe('HomePage', () => {
-  let component: HomePage;
-  let fixture: ComponentFixture<HomePage>;
+describe('DashboardPage', () => {
+  let component: DashboardPage;
+  let fixture: ComponentFixture<DashboardPage>;
   const routerSpy = jasmine.createSpyObj('Router', ['navigateByUrl']);
 
   configureTestSuite(() => {
     TestBed.configureTestingModule({
-      declarations: [HomePage],
+      declarations: [DashboardPage],
       imports: [
         RouterTestingModule.withRoutes(
           [
-            { path: '', component: HomePage },
+            { path: '', component: DashboardPage },
           ],
         ),
         IonicModule.forRoot(),
@@ -37,7 +37,7 @@ describe('HomePage', () => {
   });
 
   beforeEach(async(() => {
-    fixture = TestBed.createComponent(HomePage);
+    fixture = TestBed.createComponent(DashboardPage);
     component = fixture.componentInstance;
     fixture.detectChanges();
   }));
