+++ src/app/pages/dashboard/dashboard.page.ts
@@ -3,16 +3,16 @@ import { Router } from '@angular/router';
 import { Store } from '@ngrx/store';
 import { AlertController, Platform } from '@ionic/angular';
 import { Observable } from 'rxjs';
-import { BasePageComponent } from '../shared/classes/base-page';
-import { AuthenticationProvider } from '../providers/authentication/authentication';
-import { StoreModel } from '../shared/models/store.model';
-import { selectEmployeeName, selectVersionNumber } from '../../store/app-info/app-info.selectors';
+import { BasePageComponent } from '../../shared/classes/base-page';
+import { AuthenticationProvider } from '../../providers/authentication/authentication';
+import { StoreModel } from '../../../types/store.model';
+import { selectEmployeeName, selectVersionNumber } from '../../../store/app-info/app-info.selectors';
 @Component({
-  selector: 'app-home',
-  templateUrl: 'home.page.html',
-  styleUrls: ['home.page.scss'],
+  selector: 'app-dashboard',
+  templateUrl: 'dashboard.page.html',
+  styleUrls: ['dashboard.page.scss'],
 })
-export class HomePage extends BasePageComponent {
+export class DashboardPage extends BasePageComponent {
 
   appVersion$: Observable<string>;
   employeeName$: Observable<string>;
