+++ src/app/pages/login/login.page.html
@@ -7,9 +7,9 @@
 <ion-content>
 
 <div i>
-<ion-button color="dark" name="login-button" (click)="goToHome()">
+<ion-button color="dark" name="login-button" (click)="goToDashboard()">
 <ion-icon name="log-in-outline"></ion-icon>
-home
+Dashboard
 </ion-button>
 </div>
